updated mail class and code documentation
