+++ doc/UML_Class.uxf
@@ -1,0 +1,23 @@
+<diagram program="umletino" version="15.0.0"><zoom_level>10</zoom_level><element><id>UMLClass</id><coordinates><x>120</x><y>60</y><w>260</w><h>280</h></coordinates><panel_attributes>mail
+Package::-
+--
+-state: enum
+-config: dict
+
+--
++configure(config: dict): bool
++send(config: dict): none
+
+-readAttachments(path:string): string (base64)
+-createHtmlBody(config:dict): string
+
+-sendThread(config: dict): none
+--
+
+Responsibilities
+-- Resp1
+*-- Resp2*</panel_attributes><additional_attributes></additional_attributes></element><element><id>UMLClass</id><coordinates><x>90</x><y>500</y><w>210</w><h>70</h></coordinates><panel_attributes>_object: mail_
+--
+state: enum="idle"
+[waiting for message]</panel_attributes><additional_attributes></additional_attributes></element><element><id>Relation</id><coordinates><x>180</x><y>330</y><w>120</w><h>190</h></coordinates><panel_attributes>lt=&lt;.
+&lt;&lt;instanceOf&gt;&gt;</panel_attributes><additional_attributes>40;10;10;170</additional_attributes></element></diagram>
