+++ doc/mail/sequence.uxf
@@ -1,0 +1,9 @@
+<diagram program="umletino" version="15.0.0"><zoom_level>19</zoom_level><help_text>Space for diagram notes</help_text><element><id>UMLFrame</id><coordinates><x>0</x><y>0</y><w>1121</w><h>1824</h></coordinates><panel_attributes>Mail</panel_attributes><additional_attributes></additional_attributes></element><element><id>UMLGeneric</id><coordinates><x>95</x><y>76</y><w>190</w><h>57</h></coordinates><panel_attributes>_Report:Instance_</panel_attributes><additional_attributes></additional_attributes></element><element><id>Relation</id><coordinates><x>171</x><y>114</y><w>57</w><h>1653</h></coordinates><panel_attributes>lt=.</panel_attributes><additional_attributes>10;10;10;850</additional_attributes></element><element><id>UMLGeneric</id><coordinates><x>399</x><y>76</y><w>190</w><h>57</h></coordinates><panel_attributes>_mail:Instance_</panel_attributes><additional_attributes></additional_attributes></element><element><id>UMLGeneric</id><coordinates><x>703</x><y>76</y><w>304</w><h>57</h></coordinates><panel_attributes>_eliona handler:Instance_</panel_attributes><additional_attributes></additional_attributes></element><element><id>Relation</id><coordinates><x>475</x><y>114</y><w>57</w><h>1653</h></coordinates><panel_attributes>lt=.</panel_attributes><additional_attributes>10;10;10;850</additional_attributes></element><element><id>Relation</id><coordinates><x>836</x><y>114</y><w>57</w><h>1425</h></coordinates><panel_attributes>lt=.</panel_attributes><additional_attributes>10;10;10;730</additional_attributes></element><element><id>UMLGeneric</id><coordinates><x>475</x><y>266</y><w>38</w><h>209</h></coordinates><panel_attributes>bg=white</panel_attributes><additional_attributes></additional_attributes></element><element><id>Relation</id><coordinates><x>171</x><y>228</y><w>342</w><h>76</h></coordinates><panel_attributes>lt=&lt;&lt;-
+send</panel_attributes><additional_attributes>160;20;10;20</additional_attributes></element><element><id>Relation</id><coordinates><x>171</x><y>437</y><w>342</w><h>76</h></coordinates><panel_attributes>lt=&lt;&lt;-
+config/data Feedback</panel_attributes><additional_attributes>10;20;160;20</additional_attributes></element><element><id>UMLGeneric</id><coordinates><x>475</x><y>551</y><w>38</w><h>209</h></coordinates><panel_attributes>bg=white</panel_attributes><additional_attributes></additional_attributes></element><element><id>UMLGeneric</id><coordinates><x>836</x><y>551</y><w>38</w><h>209</h></coordinates><panel_attributes>bg=white</panel_attributes><additional_attributes></additional_attributes></element><element><id>Relation</id><coordinates><x>494</x><y>513</y><w>380</w><h>76</h></coordinates><panel_attributes>lt=&lt;&lt;-
+post_mail</panel_attributes><additional_attributes>180;20;10;20</additional_attributes></element><element><id>Relation</id><coordinates><x>494</x><y>722</y><w>380</w><h>76</h></coordinates><panel_attributes>lt=&lt;&lt;-
+mailId:int</panel_attributes><additional_attributes>10;20;180;20</additional_attributes></element><element><id>UMLFrame</id><coordinates><x>399</x><y>1159</y><w>551</w><h>304</h></coordinates><panel_attributes>loop</panel_attributes><additional_attributes></additional_attributes></element><element><id>Relation</id><coordinates><x>494</x><y>1178</y><w>380</w><h>95</h></coordinates><panel_attributes>lt=&lt;&lt;-
+request_mail_state
+(maiId:int)</panel_attributes><additional_attributes>180;20;10;20</additional_attributes></element><element><id>UMLGeneric</id><coordinates><x>475</x><y>1216</y><w>38</w><h>209</h></coordinates><panel_attributes>bg=white</panel_attributes><additional_attributes></additional_attributes></element><element><id>Relation</id><coordinates><x>494</x><y>1387</y><w>380</w><h>76</h></coordinates><panel_attributes>lt=&lt;&lt;-
+mailState:string</panel_attributes><additional_attributes>10;20;180;20</additional_attributes></element><element><id>UMLGeneric</id><coordinates><x>836</x><y>1216</y><w>38</w><h>209</h></coordinates><panel_attributes>bg=white</panel_attributes><additional_attributes></additional_attributes></element><element><id>UMLGeneric</id><coordinates><x>475</x><y>912</y><w>38</w><h>114</h></coordinates><panel_attributes></panel_attributes><additional_attributes></additional_attributes></element><element><id>UMLFrame</id><coordinates><x>399</x><y>836</y><w>551</w><h>247</h></coordinates><panel_attributes>wait TimeScheduled reached</panel_attributes><additional_attributes></additional_attributes></element><element><id>UMLGeneric</id><coordinates><x>475</x><y>1539</y><w>38</w><h>114</h></coordinates><panel_attributes></panel_attributes><additional_attributes></additional_attributes></element><element><id>Relation</id><coordinates><x>171</x><y>1558</y><w>342</w><h>76</h></coordinates><panel_attributes>lt=&lt;-
+msgState:enum</panel_attributes><additional_attributes>10;20;160;20</additional_attributes></element><element><id>UMLSpecialState</id><coordinates><x>836</x><y>1482</y><w>38</w><h>38</h></coordinates><panel_attributes>type=termination</panel_attributes><additional_attributes></additional_attributes></element></diagram>
